+++ src/components/Hotel/Hotel/Styles.js
